video : merge puppet & viewer & refactor control
