+++ libs/media/src/lib/file/controls/controls.component.ts
@@ -1,7 +1,8 @@
 import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 
 import { Event, EventService } from '@blockframes/event/+state';
-import { Meeting, MeetingPdfControl } from '@blockframes/event/+state/event.firestore';
+import { Meeting, MeetingMediaControl } from '@blockframes/event/+state/event.firestore';
+import { debounceFactory } from '@blockframes/utils/helpers';
 
 @Component({
   selector: '[event] file-controls',
@@ -13,11 +14,14 @@ export class FileControlsComponent {
 
   @Input() event: Event<Meeting>;
 
+  /** A debounced version of `updateRemoteControl` to avoid writing on the db more than every 1s */
+  private debouncedUpdateRemoteControl = debounceFactory((newControl: MeetingMediaControl) => this.updateRemoteControl(newControl), 1000);
+
   constructor(
     private eventService: EventService,
   ) { }
 
-  controlChange(control: MeetingPdfControl) {
+  updateRemoteControl(control: MeetingMediaControl) {
     this.eventService.update({
       ...this.event,
       meta: {
@@ -28,5 +32,9 @@ export class FileControlsComponent {
         },
       }
     });
+  }
+
+  controlChange(control: MeetingMediaControl) {
+    this.debouncedUpdateRemoteControl(control);
   }
 }
