+++ libs/media/src/lib/file/puppets/puppets.component.scss
@@ -29,7 +29,7 @@ section {
 
 section.owner {
   pdf-viewer,
-  video-puppet,
+  video-viewer,
   img {
     max-height: 85%;
   }
