--- libs/media/src/lib/video/puppet/puppet.component.html
@@ -1,25 +1,0 @@
-<section fxLayout="column" fxLayoutAlign="center center" #container>
-
-  <div id="puppet-player" >
-    <!-- everything in the div will be destroyed when the player is loaded -->
-    <section fxLayout="column">
-      <ng-container *ngIf="loading$ | async; else error">
-        <mat-spinner></mat-spinner>
-      </ng-container>
-      <ng-template #error>
-        <img asset="add_organization.webp" />
-        <p>Ooops ! Sorry, something broke.</p>
-        <p>Please contact us if you keep seeing this error.</p>
-      </ng-template>
-    </section>
-  </div>
-
-  <button *ngIf="!fullScreen" fxFlexAlign="end" (click)="toggleFullScreen()" mat-icon-button>
-    <mat-icon svgIcon="fullscreen"></mat-icon>
-  </button>
-
-  <button *ngIf="fullScreen" fxFlexAlign="end" (click)="toggleFullScreen()" class="fullscreen" mat-icon-button>
-    <mat-icon svgIcon="fullscreen_exit"></mat-icon>
-  </button>
-
-</section>
