+++ libs/media/src/lib/video/viewer/viewer.component.html
@@ -1,17 +1,33 @@
-<div [id]="playerContainerId" class="container">
-  <!-- everything in the div will be destroyed when the player is loaded -->
-  <section fxLayout="column" fxLayoutAlign="center center">
-    <ng-container *ngIf="loading$ | async; else error">
-      <mat-spinner></mat-spinner>
-    </ng-container>
-    <ng-template #error>
-      <img asset="add_organization.webp" />
-      <p>Ooops ! Sorry, something broke.</p>
-      <p>Please contact us if you keep seeing this error.</p>
-      <button mat-flat-button (click)="refresh()">
-        <mat-icon svgIcon="refresh"></mat-icon>
-        <span>Retry</span>
-      </button>
-    </ng-template>
-  </section>
-</div>
+<section fxLayout="column" fxLayoutAlign="center center" #container>
+
+  <div [id]="playerContainerId">
+    <!-- everything in the div will be destroyed when the player is loaded -->
+    <section fxLayout="column" fxLayoutAlign="center center">
+      <ng-container *ngIf="loading$ | async; else error">
+        <mat-spinner></mat-spinner>
+      </ng-container>
+      <ng-template #error>
+        <img asset="add_organization.webp" />
+        <p>Ooops ! Sorry, something broke.</p>
+        <p>Please contact us if you keep seeing this error.</p>
+
+        <button mat-flat-button (click)="refresh()">
+          <mat-icon svgIcon="refresh"></mat-icon>
+          <span>Retry</span>
+        </button>
+      </ng-template>
+    </section>
+  </div>
+
+
+  <ng-container *ngIf="!!control">
+    <button *ngIf="!fullScreen" fxFlexAlign="end" (click)="toggleFullScreen()" mat-icon-button>
+      <mat-icon svgIcon="fullscreen"></mat-icon>
+    </button>
+
+    <button *ngIf="fullScreen" fxFlexAlign="end" (click)="toggleFullScreen()" class="fullscreen" mat-icon-button>
+      <mat-icon svgIcon="fullscreen_exit"></mat-icon>
+    </button>
+  </ng-container>
+
+</section>
