+++ libs/media/src/lib/video/viewer/viewer.component.scss
@@ -6,37 +6,45 @@
  */
  video-viewer {
   height: 100%;
-  div.jwplayer {
-    height: 100%;
-    section {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      height: 100%;
-    }
 
-    div {
-      div.jw-logo.jw-logo-top-right.jw-reset {
-        /* We need to use `!important` because jwplayer style is inline */
-        width: 100% !important;
-        height: 100% !important;
-        margin: 0 !important;
-        background-position: 80% 25%;
-        background-size: 50%;
-        pointer-events: none;
+  section {
+    height: 100%;
+
+    div.jwplayer {
+      height: 100%;
+
+      section {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        height: 100%;
+      }
+
+      div {
+        div.jw-logo.jw-logo-top-right.jw-reset {
+          /* We need to use `!important` because jwplayer style is inline */
+          width: 100% !important;
+          height: 100% !important;
+          margin: 0 !important;
+          background-position: 80% 25%;
+          background-size: 50%;
+          pointer-events: none;
+        }
       }
     }
   }
 
-  .container {
-    height: 100%;
-    section {
-      height: 100%;
-      width: 100%;
-    }
+  p {
+    text-align: center;
+  }
+  img {
+    width: 100px;
   }
 
-  img {
-    width: 200px;
+  button.fullscreen {
+    position: absolute;
+    bottom: 0;
+    background: #000;
   }
+
 }
