+++ libs/media/src/lib/video/viewer/viewer.module.ts
@@ -1,25 +1,30 @@
+
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { VideoViewerComponent } from './viewer.component';
-import { FlexLayoutModule } from '@angular/flex-layout';
-import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
-import { ImageModule } from '@blockframes/media/image/directives/image.module';
-import { MatButtonModule } from '@angular/material/button';
+
 import { MatIconModule } from '@angular/material/icon';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatButtonModule } from '@angular/material/button';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
+
+import { ImageModule } from '@blockframes/media/image/directives/image.module';
+
+import { VideoViewerComponent } from '../../video/viewer/viewer.component';
 
 @NgModule({
   declarations: [ VideoViewerComponent ],
-  exports: [ VideoViewerComponent ],
   imports: [
     CommonModule,
-    FlexLayoutModule,
-    ImageModule,
 
-    MatIconModule,
-    MatButtonModule,
-    MatSnackBarModule,
+    FlexLayoutModule,
     MatProgressSpinnerModule,
-  ]
+    MatButtonModule,
+    MatIconModule,
+    MatSnackBarModule,
+
+    ImageModule,
+  ],
+  exports: [ VideoViewerComponent ],
 })
-export class VideoViewerModule { }
+export class VideoViewerModule {}
